Faas 4: tõsta autentimine ja registreerimine routerisse #36
